+++ scripts/utils/helpers.js
@@ -20,14 +20,19 @@ const mv = (src, dst, { story = mainStory } = {}) => {
   shell.mv('-rf', path.normalize(src), path.normalize(dst));
 };
 
-const exec = async (cmd, { story = mainStory, logLevel = 'info', cwd } = {}) => {
+const exec = async (cmd, {
+  story = mainStory,
+  logLevel = 'info',
+  errorLogLevel = 'error',
+  cwd,
+} = {}) => {
   const prevWd = shell.pwd();
   let title = `Run cmd ${chalk.green.bold(cmd)}`;
   if (cwd) title += ` at ${chalk.green(cwd)}`;
   const ownStory = story.child({ title, level: logLevel });
   try {
     if (cwd) cd(cwd, { story });
-    const result = await _exec(cmd, ownStory);
+    const result = await _exec(cmd, { story: ownStory, errorLogLevel });
     if (cwd) cd(prevWd, { story });
     return result;
   } finally {
@@ -35,10 +40,10 @@ const exec = async (cmd, { story = mainStory, logLevel = 'info', cwd } = {}) => 
   }
 };
 
-const _exec = (cmd, story) => new Promise((resolve, reject) => {
+const _exec = (cmd, { story, errorLogLevel }) => new Promise((resolve, reject) => {
   const child = shell.exec(cmd, { silent: true }, (code, stdout, stderr) => {
     if (code !== 0) {
-      story.error(`Command failed [${code}]`);
+      story[errorLogLevel](`Command failed [${code}]`);
       reject(new Error(`Command failed: ${cmd}`));
       return;
     }
@@ -50,7 +55,7 @@ const _exec = (cmd, story) => new Promise((resolve, reject) => {
     story.info(cmdName, `| ${line}`);
   });
   child.stderr.pipe(split()).on('data', (line) => {
-    if (line) story.error(cmdName, `| ${line}`);
+    if (line) story[errorLogLevel](cmdName, `| ${line}`);
   });
 });